fase_1/tech_challenge/toggle-master-monolith/requirements.txt
Flask==2.2.2
Werkzeug==2.3.8
psycopg2-binary==2.9.5
gunicorn==20.1.0
boto3